Exclude "underscored" `.proto` files from the build.
This change filters out any `.proto` files that are in underscore-prefixed
directories, or themselves start with an underscore.

By convention, under Go, any such `.go` files are exluded from the build. As
such it is common to rename a directory to have an underscore prefix while
performing large refactories or rewrites. It is convenient if any `.proto`
files under such directories are also ignored, as they often do not compile
properly within the underscore-prefixed directory anyway.