Add support for building Go plugins.
If a binary is built with the extension `.so` or `.dll` it is built as a plugin
rather than a regular executable.

Building plugins appears to (logically) require CGO to be enabled. This change
does NOT automatically enable CGO when building a plugin, the user is still
required to explicitly enable set `CGO_ENABLED = 1` at the top of their
project's makefile.