Change `clean::` to `clean:` to prevent ordering issues.
This changes the `clean` target to dissallow additional recipes in
language-specific makefiles, and instead builds a `_clean` target as the first
step of `clean`.

Language-specific makefiles that have cleanup logic will need to define a
`_clean::` target instead of `clean::`. This ensures we can run the
language-specific logic before the makefiles themselves are removed.