Remove helm-lint from lint and precommit targets
We're doing this because helm lint requires the dependencies to be
installed, and it's often difficult to get the dependencies installed
on developer's machines when there are private chart dependencies,
because it requires short-lived ECR auth.

Longer term, it would be nice to "tolerate" helm lint failures on
developer's machines when the deps can't be installed.